refactoring in preparation for imagestream restore action
Moves functions that are needed by both restore and backup actions
for a plugin into shared.go, in the same package as backup/restore.

Also, moves functions that are needed by multiple plugins into the
common package, except for the clients (CoreClient, ImageClient,
etc), which are moved into the clients package.

The latter part of this should resolve
https://github.com/fusor/ocp-velero-plugin/issues/8
