+++ velero-plugins/imagestream/shared.go
@@ -1,0 +1,65 @@
+package imagestream
+
+import (
+	"context"
+	"fmt"
+
+	"github.com/containers/image/copy"
+	"github.com/containers/image/signature"
+	"github.com/containers/image/transports/alltransports"
+	"github.com/containers/image/types"
+
+	"k8s.io/client-go/rest"
+)
+
+func copyImage(src, dest string, sourceCtx, destinationCtx *types.SystemContext) (string, error) {
+	policyContext, err := getPolicyContext()
+	if err != nil {
+		return "", fmt.Errorf("Error loading trust policy: %v", err)
+	}
+	defer policyContext.Destroy()
+
+	srcRef, err := alltransports.ParseImageName(src)
+	if err != nil {
+		return "", fmt.Errorf("Invalid source name %s: %v", src, err)
+	}
+	destRef, err := alltransports.ParseImageName(dest)
+	if err != nil {
+		return "", fmt.Errorf("Invalid destination name %s: %v", dest, err)
+	}
+	manifest, err := copy.Image(context.Background(), policyContext, destRef, srcRef, &copy.Options{
+		SourceCtx:      sourceCtx,
+		DestinationCtx: destinationCtx,
+	})
+	return string(manifest), err
+}
+
+func getPolicyContext() (*signature.PolicyContext, error) {
+	policy := &signature.Policy{Default: []signature.PolicyRequirement{signature.NewPRInsecureAcceptAnything()}}
+	return signature.NewPolicyContext(policy)
+}
+
+func internalRegistrySystemContext() (*types.SystemContext, error) {
+	config, err := rest.InClusterConfig()
+	if err != nil {
+		return nil, err
+	}
+
+	ctx := &types.SystemContext{
+		DockerDaemonInsecureSkipTLSVerify: true,
+		DockerInsecureSkipTLSVerify:       types.OptionalBoolTrue,
+		DockerAuthConfig: &types.DockerAuthConfig{
+			Username: "ignored",
+			Password: config.BearerToken,
+		},
+	}
+	return ctx, nil
+}
+
+func migrationRegistrySystemContext() (*types.SystemContext, error) {
+	ctx := &types.SystemContext{
+		DockerDaemonInsecureSkipTLSVerify: true,
+		DockerInsecureSkipTLSVerify:       types.OptionalBoolTrue,
+	}
+	return ctx, nil
+}
