+++ velero-plugins/route/backup.go
@@ -4,12 +4,10 @@ import (
 	"github.com/sirupsen/logrus"
 
 	//routev1API "github.com/openshift/api/build/v1"
-	routev1 "github.com/openshift/client-go/route/clientset/versioned/typed/route/v1"
 	"k8s.io/apimachinery/pkg/api/meta"
 
 	//metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
-	"k8s.io/client-go/rest"
 
 	v1 "github.com/heptio/velero/pkg/apis/velero/v1"
 	"github.com/heptio/velero/pkg/backup"
@@ -43,23 +41,11 @@ func (p *BackupPlugin) Execute(item runtime.Unstructured, backup *v1.Backup) (ru
 
 	annotations["openshift.io/route-plugin"] = "1"
 
-	/*client, err := p.routeClient()
+	/*client, err := clients.NewRouteClient()
 	if err != nil {
 		return nil, nil, err
 	}*/
 	metadata.SetAnnotations(annotations)
 
 	return item, nil, nil
-}
-
-func (p *BackupPlugin) routeClient() (*routev1.RouteV1Client, error) {
-	config, err := rest.InClusterConfig()
-	if err != nil {
-		return nil, err
-	}
-	client, err := routev1.NewForConfig(config)
-	if err != nil {
-		return nil, err
-	}
-	return client, nil
 }
